fix(view): reject non-positive/absurd zoom in layout set-zoom; extract pure parser
set-zoom now rejects values outside [1e-50,1e50] (Cytoscape's clamp range) at parse time with exit 2, before any daemon mutation. To stay under the 500-line limit the pure parsing layer (ADT, usage, parseViewCommand) moved to viewParse.ts and shared test lifecycle to view-test-helpers.ts; the single public runViewCommand surface is unchanged.
